refactor(storage/gcp): create mappers.go for pure data transformation logic
